+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - [TaxonKit v0.12.1](https://github.com/shenwei356/taxonkit/releases/tag/v0.12.1)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/taxonkit/v0.12.1/total.svg)](https://github.com/shenwei356/taxonkit/releases/tag/v0.12.1)
+    - `taxonkit reformat`: add a new placeholder `{K}` for rank `kingdom`. [#64](https://github.com/shenwei356/taxonkit/issues/64)
     - `taxonkit create-taxdump`: 
         - fix merged.dmp and delnodes.dmp. [gtdb-taxdump/issues/2](https://github.com/shenwei356/gtdb-taxdump/issues/2)
         - fix bug of handling non-GTDB data when using `-A/--field-accession` and no rank names given.